🚚 move http logs to its own file
