+++ frontend/app/routes/deployments/deployment-http-logs.tsx
@@ -1,31 +1,11 @@
 import { useInfiniteQuery } from "@tanstack/react-query";
-import {
-  CheckIcon,
-  ChevronRightIcon,
-  ChevronsUpDownIcon,
-  Code,
-  CopyIcon,
-  FilterIcon,
-  LoaderIcon
-} from "lucide-react";
+import { ChevronsUpDownIcon, LoaderIcon } from "lucide-react";
 import * as React from "react";
 import { useSearchParams } from "react-router";
 import { type Writeable, z } from "zod";
-import { CopyButton } from "~/components/copy-button";
-import { StatusBadge, type StatusBadgeColor } from "~/components/status-badge";
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger
-} from "~/components/ui/accordion";
+import { HttpLogRequestDetails } from "~/components/http-log-request-details";
 import { Button } from "~/components/ui/button";
-import {
-  Sheet,
-  SheetContent,
-  SheetHeader,
-  SheetTitle
-} from "~/components/ui/sheet";
+
 import {
   Table,
   TableBody,
@@ -49,7 +29,6 @@ import {
 } from "~/lib/queries";
 import { cn, formatLogTime } from "~/lib/utils";
 import { queryClient } from "~/root";
-import { formattedTime, pluralize, wait } from "~/utils";
 import { type Route } from "./+types/deployment-http-logs";
 
 export async function clientLoader({
@@ -71,6 +50,7 @@ export async function clientLoader({
     request_host: search.request_host,
     request_ip: search.request_ip,
     request_path: search.request_path,
+    request_user_agent: search.request_user_agent,
     status: search.status
   } satisfies DeploymentHTTPLogFilters;
 
@@ -119,6 +99,7 @@ export default function DeploymentHttpLogsPage({
     request_host: search.request_host,
     request_ip: search.request_ip,
     request_path: search.request_path,
+    request_user_agent: search.request_user_agent,
     status: search.status
   } satisfies DeploymentHTTPLogFilters;
 
@@ -138,7 +119,7 @@ export default function DeploymentHttpLogsPage({
 
   return (
     <>
-      <LogRequestDetails
+      <HttpLogRequestDetails
         open={Boolean(loaderData.httpLog)}
         log={loaderData.httpLog}
         onClose={() => {
@@ -290,339 +271,6 @@ function LogTableRow({ log, onClick }: LogTableRowProps) {
       </TableCell>
     </TableRow>
   );
-}
-
-type LogRequestDetailsProps = {
-  log?: HttpLog;
-  open?: boolean;
-  onClose?: () => void;
-};
-
-export function LogRequestDetails({
-  log,
-  onClose,
-  open = false
-}: LogRequestDetailsProps) {
-  return (
-    <Sheet
-      open={open}
-      onOpenChange={(open) => {
-        if (!open) {
-          onClose?.();
-        }
-      }}
-    >
-      <SheetContent
-        side="right"
-        className="z-99 border-border flex flex-col gap-4 overflow-y-auto"
-      >
-        {log && <LogRequestDetailsContent log={log} />}
-      </SheetContent>
-    </Sheet>
-  );
-}
-
-function LogRequestDetailsContent({ log }: { log: HttpLog }) {
-  const searchParams = new URLSearchParams(log.request_query ?? "");
-  return (
-    <>
-      <SheetHeader>
-        <SheetTitle className="font-normal text-card-foreground mt-5 text-base">
-          <span className="border border-gray-600 bg-gray-600/10 px-2 py-1 border-opacity-60 rounded-md ">
-            {log.request_method}
-          </span>
-          &nbsp;
-          <span className="font-medium">{log.request_path}</span>
-        </SheetTitle>
-      </SheetHeader>
-      <hr className="border-border -mx-6" />
-      <h3>Request metadata:</h3>
-
-      <dl className="flex flex-col gap-x-4 gap-y-2 items-center auto-rows-max">
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full">
-          <dt className="text-grey  inline-flex items-center">ID</dt>
-          <dd className="text-sm">{log.request_id}</dd>
-        </div>
-
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full">
-          <dt className="text-grey  inline-flex items-center">Status code</dt>
-          <dd
-            className={cn("", {
-              "text-blue-600": log.status.toString().startsWith("1"),
-              "text-green-600": log.status.toString().startsWith("2"),
-              "text-grey": log.status.toString().startsWith("3"),
-              "text-yellow-600": log.status.toString().startsWith("4"),
-              "text-red-600": log.status.toString().startsWith("5")
-            })}
-          >
-            {log.status}
-          </dd>
-        </div>
-
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full">
-          <dt className="text-grey  inline-flex items-center">Date</dt>
-          <dd className="text-sm">
-            <time
-              className="text-grey whitespace-nowrap"
-              dateTime={new Date(log.time).toISOString()}
-            >
-              {formattedTime(log.time)}
-            </time>
-          </dd>
-        </div>
-
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full">
-          <dt className="text-grey  inline-flex items-center">Duration</dt>
-          <dd className="text-sm">
-            {Intl.NumberFormat("en-US").format(
-              log.request_duration_ns / 1_000_000
-            )}
-            <span className="text-grey">ms</span>
-          </dd>
-        </div>
-
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full">
-          <dt className="text-grey  inline-flex items-center">Protocol</dt>
-          <dd className="text-sm">{log.request_protocol}</dd>
-        </div>
-
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full group">
-          <dt className="text-grey inline-flex items-center">
-            <span>Client IP</span>
-            <TooltipProvider>
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <Button
-                    variant="ghost"
-                    className="px-2.5 py-0.5 md:opacity-0 focus-visible:opacity-100 group-hover:opacity-100"
-                  >
-                    <FilterIcon size={15} />
-                    <span className="sr-only">Filter this IP</span>
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>Filter this IP</TooltipContent>
-              </Tooltip>
-
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <CopyButton label="Copy value" value={log.request_ip} />
-                </TooltipTrigger>
-                <TooltipContent>Copy value</TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          </dt>
-          <dd className="text-sm break-all text-grey">{log.request_ip}</dd>
-        </div>
-
-        <div className="grid grid-cols-2 items-start gap-x-4 w-full group">
-          <dt className="text-grey inline-flex items-center gap-1">
-            <span>User Agent</span>
-            <TooltipProvider>
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <Button
-                    variant="ghost"
-                    className="px-2.5 py-0.5 md:opacity-0 focus-visible:opacity-100 group-hover:opacity-100"
-                  >
-                    <FilterIcon size={15} />
-                    <span className="sr-only">Filter this user-agent</span>
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>Filter this user-agent</TooltipContent>
-              </Tooltip>
-
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <CopyButton
-                    label="Copy value"
-                    value={log.request_user_agent ?? ""}
-                  />
-                </TooltipTrigger>
-                <TooltipContent>Copy value</TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          </dt>
-          <dd className="text-sm text-grey">
-            {log.request_user_agent ?? <span className="font-mono">N/A</span>}
-          </dd>
-        </div>
-      </dl>
-
-      <hr className="border-border -mx-6" />
-
-      <h3>URL data:</h3>
-      <dl className="flex flex-col gap-x-4 gap-y-2 items-center auto-rows-max">
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full group">
-          <dt className="text-grey inline-flex items-center gap-1 group">
-            <span>Host</span>
-            <TooltipProvider>
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <Button
-                    variant="ghost"
-                    className="px-2.5 py-0.5 md:opacity-0 focus-visible:opacity-100 group-hover:opacity-100"
-                  >
-                    <FilterIcon size={15} />
-                    <span className="sr-only">Filter this host</span>
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>Filter this host</TooltipContent>
-              </Tooltip>
-
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <CopyButton label="Copy value" value={log.request_host} />
-                </TooltipTrigger>
-                <TooltipContent>Copy value</TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          </dt>
-          <dd className="text-sm">{log.request_host}</dd>
-        </div>
-
-        <div className="grid grid-cols-2 items-center gap-x-4 w-full">
-          <dt className="text-grey inline-flex items-center gap-1 group">
-            <span>Pathname</span>
-            <TooltipProvider>
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <Button
-                    variant="ghost"
-                    className="px-2.5 py-0.5 md:opacity-0 focus-visible:opacity-100 group-hover:opacity-100"
-                  >
-                    <FilterIcon size={15} />
-                    <span className="sr-only">Filter this pathname</span>
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>Filter this pathname</TooltipContent>
-              </Tooltip>
-
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger asChild>
-                  <CopyButton label="Copy value" value={log.request_path} />
-                </TooltipTrigger>
-                <TooltipContent>Copy value</TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          </dt>
-          <dd className="text-sm">{log.request_path}</dd>
-        </div>
-
-        {log.request_query && (
-          <div className="grid grid-cols-2 items-center gap-x-4 w-full border-b-0 border-border pb-2">
-            <dt className="text-grey inline-flex items-center gap-1 group">
-              <span>Query</span>
-              <TooltipProvider>
-                <Tooltip delayDuration={0}>
-                  <TooltipTrigger asChild>
-                    <Button
-                      variant="ghost"
-                      className="px-2.5 py-0.5 md:opacity-0 focus-visible:opacity-100 group-hover:opacity-100"
-                    >
-                      <FilterIcon size={15} />
-                      <span className="sr-only">Filter this query</span>
-                    </Button>
-                  </TooltipTrigger>
-                  <TooltipContent>Filter this query</TooltipContent>
-                </Tooltip>
-
-                <Tooltip delayDuration={0}>
-                  <TooltipTrigger asChild>
-                    <CopyButton
-                      label="Copy value"
-                      value={`?${log.request_query}`}
-                    />
-                  </TooltipTrigger>
-                  <TooltipContent>Copy value</TooltipContent>
-                </Tooltip>
-              </TooltipProvider>
-            </dt>
-            <dd className="text-sm">
-              <span className="text-grey">{"?"}</span>
-              {searchParams.entries().map(([key, value], index) => (
-                <span key={`${key}-${index}`}>
-                  <span className="text-link">{key}</span>
-                  {value && (
-                    <>
-                      <span className="text-grey">{"="}</span>
-                      <span className="text-card-foreground break-all">
-                        {value}
-                      </span>
-                    </>
-                  )}
-                  {index < searchParams.size - 1 && (
-                    <span className="text-grey">{"&"}</span>
-                  )}
-                </span>
-              ))}
-            </dd>
-          </div>
-        )}
-      </dl>
-
-      <hr className="border-border -mx-6" />
-      <Accordion type="single" collapsible defaultValue="request">
-        <AccordionItem value="request" className="border-0">
-          <AccordionTrigger className="gap-2 py-0 data-[state=open]:pb-4">
-            <ChevronRightIcon className="h-4 w-4 shrink-0 transition-transform duration-200" />
-            <h3>Response headers:</h3>
-          </AccordionTrigger>
-          <AccordionContent className="flex flex-col pb-0">
-            <dl className="flex flex-col gap-0.5 text-base">
-              {transformHeadersObjectToArray(log.response_headers).map(
-                ([key, value]) => (
-                  <div className="inline gap-1 w-full" key={`${key}-${value}`}>
-                    <dt className="text-link inline flex-none">{key}:</dt>
-                    &nbsp;&nbsp;
-                    <dd className="break-all inline text-grey">{value}</dd>
-                  </div>
-                )
-              )}
-            </dl>
-          </AccordionContent>
-        </AccordionItem>
-      </Accordion>
-      <hr className="border-border -mx-6" />
-      <Accordion type="single" collapsible defaultValue="response">
-        <AccordionItem value="response" className="border-0">
-          <AccordionTrigger className="gap-2 py-0 data-[state=open]:pb-4">
-            <ChevronRightIcon className="h-4 w-4 shrink-0 transition-transform duration-200" />
-            <h3>Request headers:</h3>
-          </AccordionTrigger>
-          <AccordionContent className="flex flex-col pb-0">
-            <dl className="flex flex-col gap-0.5 text-base">
-              {transformHeadersObjectToArray(log.request_headers).map(
-                ([key, value]) => (
-                  <div className="inline gap-1 w-full" key={`${key}-${value}`}>
-                    <dt className="text-link inline flex-none">{key}:</dt>
-                    &nbsp;&nbsp;
-                    <dd className="break-all inline text-grey">{value}</dd>
-                  </div>
-                )
-              )}
-            </dl>
-          </AccordionContent>
-        </AccordionItem>
-      </Accordion>
-      <hr className="border-border -mx-6" />
-    </>
-  );
-}
-
-function transformHeadersObjectToArray(current: {
-  [key: string]: string[];
-}): Array<[string, string]> {
-  const target: Array<[string, string]> = [];
-
-  for (const key in current) {
-    if (current.hasOwnProperty(key)) {
-      current[key].forEach((value) => {
-        target.push([key, value]);
-      });
-    }
-  }
-
-  return target;
 }
 
 function HeaderSection() {
